refactor(adk): extract GCS signing logic into gcs_signing.py
